#164 Use retry-able connections for Hive JDBC query execution.
