+++ pramen/core/src/main/scala/za/co/absa/pramen/core/reader/TableReaderJdbc.scala
@@ -45,7 +45,7 @@ class TableReaderJdbc(tableName: String,
 
   private val extraOptions = ConfigUtils.getExtraOptions(conf, "option")
 
-  private val jdbcUrlSelector = new JdbcUrlSelector(jdbcReaderConfig.jdbcConfig)
+  private val jdbcUrlSelector = JdbcUrlSelector(jdbcReaderConfig.jdbcConfig)
 
   private val numberOfRetries = jdbcReaderConfig.connectionRetries.getOrElse(jdbcUrlSelector.getNumberOfUrls - 1)
